requirements.txt
discord.py
aiomysql
python-dotenv
yt-dlp
